add immediate in-process environment updates for CMD shell
* for CMD shells, during install or update, pass environment changes back up the call
  stack into the calling CMD shell process
* adds feature parity with the PowerShell shell to the CMD shell
  - allows immediate use of newly installed packages
  - starting a new CMD shell is *no longer required* prior to use of an
    installed/updated application
  - eg, `scoop install xxx` can be immediately followed by using `xxx`
* uses an environment variable pipe

* fixes #485

* note: variables set within the adjunctive install script are not captured unless the
  install script uses the env() subroutine
* note: due to increased code complexity, proxy execution is now required for safe CMD
  shim updating

.# Discussion

When running `scoop install ...` (or `scoop update ...`), environment changes are
performed and persisted to the USER or MACHINE store. Within a PowerShell shell, the
PROCESS environment is also updated, allowing for immediate use of the installed
application (from the same shell command line). But, when used from the CMD shell,
`scoop install ...` does not update the PROCESS environment. Because of this, a new CMD
shell must be executed to finalize the environment changes and allow use of the
installed application.

However, BAT/CMD scripts can persist environment changes to the CMD shell PROCESS
environment. By leveraging that ability, it is possible to pass environment updates back
up to a calling CMD process via "environment piping". To use "environment piping", a
temporary file is created and passed to its child process. The child process adds the
environment changes to the temporary file and then exits, returning control to the parent
process. The parent process then processes the temporary file, adding the changes to the
current CMD shell PROCESS environment.

This does mean that the parent process (the `scoop.cmd` shim) must execute significant
code after the child process (the main scoop application) returns. This impacts upon the
method used to allow safe modification of the scoop CMD shim during a `scoop update`
(see Issue #481 and the solution in commit c86cb30ec2b851967c9d44aa88202b3bf64bda8f). It
is difficult, if not impossible, to construct a parenthesis-block statement which can
load the environment changes into the PROCESS environment and then exit cleanly. Because
of this, it is necessary to switch to the proxy execution method mentioned in Issue #481
for safe CMD shim updates.