refactor appveyor testing to use a powershell sub-shell + add test summary display step
.# Discussion

AppVeyor can interfere with loading external assemblies, notably 'Newtonsoft.Json',
causing remote testing failures which are not seen with local testing (see [1]). In
concert with AppVeyor staff, it was discovered that testing in a powershell sub-shell
avoids the issue.

This, however, causes some issues with test script output, especially colored output via
`write-host`. So, the testing script was refactored to use the XML result output of
`pester`, which is already being generated for upload to AppVeyor, as a state repository
for use with later summary output. And, a script which interprets the XML testing results
and displays them in a colorful user-friendly format (`test\bin\summarize-results.ps1`)
was added into the testing code directory.

[1] http://help.appveyor.com/discussions/problems/3938-test-failures-on-appveyor-no-duplicated-locally