refactor 'Newtonsoft.Json' assembly loading technique (load from an in-memory copy)
.# Discussion

PowerShell's `import-module` command leaves an open-handle on any loaded assembly ".dll"
file until the shell is shut down. To enable source file self-updates, any open handles
within the source must be avoided.

Initially, the assembly was copied to an out-of-source location (within %TEMP%) before
being loaded. Although this process works, it leaves copies of the assembly in the %TEMP%
directory which are not subsequently cleaned up.

This commit alters that process to a "cleaner" behavior by loading the needed
'Newtonsoft.Json' assembly from an in-memory copy. There are no open handles within the
file system and everything is automatically cleaned up whenever the parent process
is shut-down.